Align font height, capHeight to FreeType metrics
Previously, `height` was derived from a manually chosen glyph (typically
"I", Standard Cap Height) using FontBuilder, and `lineSpacing` was used
inconsistently for layout logic as a height.

Now, with the FreeType-based system:
- `height` uses `face->size->metrics.height`, which includes the
  recommended line height with internal leading/line gap as defined by
  the font designer.
- `lineSpacing` was removed
- `GetCapHeight` uses the standard cap height + ascender to have a
  visual virtual alignment

This change standardizes font metric usage:
- Use `height` for vertical layout and line progression.
- Use `GetCapheight` in layout engines like `CGUIText` or `CGUIString`,

This ensures better alignment across fonts and consistent spacing in
multiline text rendering.

Fixes: #7962

+++ source/ps/CConsole.h
@@ -78,9 +78,9 @@ private:
 	// Lock for all state modified by InsertMessage
 	std::mutex m_Mutex;
 
-	int m_FontHeight;
-	int m_FontWidth;
-	int m_FontOffset; // distance to move up before drawing
+	float m_FontHeight;
+	float m_FontWidth;
+	float m_FontOffset; // distance to move up before drawing
 	size_t m_CharsPerPage;
 
 	float m_X;
